Refactor router imports
