+++ src/js/server/routes/index.ts
@@ -1,1 +1,13 @@
-require('router');
+import router from './router';
+import apiRouter from './api-router';
+import applyRouter from './apply-router';
+import eventRouter from './event-router';
+import hcApiRouter from './hc-api-router';
+
+export {
+  router,
+  apiRouter,
+  applyRouter,
+  eventRouter,
+  hcApiRouter
+};
